+++ src/features/Conversation/ChatList/Extras/TTS/index.tsx
@@ -1,9 +1,10 @@
 import { memo, useMemo } from 'react';
 import { Md5 } from 'ts-md5';
 
-import { settingsSelectors, useGlobalStore } from '@/store/global';
+import { useGlobalStore } from '@/store/global';
+import { settingsSelectors } from '@/store/global/selectors';
 import { useSessionStore } from '@/store/session';
-import { agentSelectors } from '@/store/session/slices/agent';
+import { agentSelectors } from '@/store/session/selectors';
 
 import FilePlayer from './FilePlayer';
 import InitPlayer, { TTSProps } from './InitPlayer';