+++ src/features/Conversation/ChatList/Messages/Function.tsx
@@ -4,7 +4,7 @@ import { Flexbox } from 'react-layout-kit';
 
 import { useChatStore } from '@/store/chat';
 import { chatSelectors } from '@/store/chat/selectors';
-import { ChatMessage } from '@/types/chatMessage';
+import { ChatMessage } from '@/types/message';
 
 import Inspector from '../Plugins/Inspector';
 import PluginRender from '../Plugins/Render';