+++ src/store/chat/store.ts
@@ -6,12 +6,12 @@ import { StateCreator } from 'zustand/vanilla';
 import { createHyperStorage } from '@/store/middleware/createHyperStorage';
 import { isDev } from '@/utils/env';
 
-import { ChatEnhanceAction, chatEnhance } from './actions/enhance';
-import { ChatMessageAction, chatMessage } from './actions/message';
-import { ChatPluginAction, chatPlugin } from './actions/plugin';
-import { ShareAction, chatShare } from './actions/share';
-import { ChatTopicAction, chatTopic } from './actions/topic';
 import { ChatStoreState, initialState } from './initialState';
+import { ChatEnhanceAction, chatEnhance } from './slices/enchance/action';
+import { ChatMessageAction, chatMessage } from './slices/message/action';
+import { ShareAction, chatShare } from './slices/share/action';
+import { ChatPluginAction, chatPlugin } from './slices/tool/action';
+import { ChatTopicAction, chatTopic } from './slices/topic/action';
 
 export interface ChatStoreAction
   extends ChatMessageAction,